backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf==1.24.2
black
numpy
Pillow
